Add initial version of 1-hop replication
In a nutshell, peers actively share information about their blocks (through the replicator) and keep an index of the blocks owned by their neighbors.
The index is used by the engine to send a new type of reply (know) which contains a cid and a list of peers that are known to have the block. Sessions leverage this information to open new connections (if needed) and request blocks. The index is also used by sessions to perform a "filtered" broadcast that instead of broadcasting all cids sends wants (have or block) to peers that possibly have the correspondent block.

Also adds a tracer to log all messages and some other logs that will be needed to perform an experimental evaluation and updates the .gitignore.